OffRag/requirements.txt
streamlit

chromadb

embedchain

pdfplumber

Pillow

pytesseract

numpy

bcrypt

requests

tenacity

matplotlib

seaborn

pandas